Set up BrowserRouter and extract LevelPage
