+++ packages/viewer/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>viewer</title>
+    <title>nandlang</title>
   </head>
   <body>
     <div id="root"></div>
